# backend/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/users/me/parcours": {
      "patch": {
        "operationId": "UserController_updateParcours",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateParcoursDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Le parcours a été mis à jour."
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Mettre à jour le parcours de l'utilisateur",
        "tags": [
          "users"
        ]
      }
    },
    "/users/me/onboarding/complete": {
      "patch": {
        "operationId": "UserController_completeOnboarding",
        "parameters": [],
        "responses": {
          "200": {
            "description": "L'onboarding a été marqué comme terminé."
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Marquer l'onboarding comme complété",
        "tags": [
          "users"
        ]
      }
    },
    "/users/me/onboarding/reset-retest": {
      "post": {
        "description": "Remet onboarding_completed à false et supprime toutes les entrées score_history de l'utilisateur (y compris quiz mensuel) pour que le prochain onboarding redevienne le « premier » bilan.",
        "operationId": "UserController_resetOnboardingForRetest",
        "parameters": [],
        "requestBody": {
          "required": false,
          "description": "Corps optionnel (ex. `{}`) pour compatibilité clients JSON.",
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "additionalProperties": true,
                "example": {}
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Onboarding réinitialisé, historique des scores effacé."
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "[TEMP] Réinitialiser onboarding et effacer les scores enregistrés (refaire le quiz)",
        "tags": [
          "users"
        ]
      }
    },
    "/users/me/simple-challenge/complete": {
      "post": {
        "description": "Crédite les feuilles de récompense du défi simple, au plus une fois par jour (UTC).",
        "operationId": "UserController_completeSimpleDailyChallenge",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Défi enregistré, feuilles mises à jour.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SimpleChallengeCompleteResponseDto"
                }
              }
            }
          },
          "409": {
            "description": "Défi déjà relevé pour ce jour UTC."
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Relever le défi quotidien (carte Accueil)",
        "tags": [
          "users"
        ]
      }
    },
    "/users/me/scores": {
      "get": {
        "operationId": "UserController_getScoreHistory",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Historique des scores récupéré avec succès",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ScoreHistoryResponseDto"
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Récupérer l'historique des scores carbone de l'utilisateur",
        "tags": [
          "users"
        ]
      }
    },
...
  },
  "info": {
    "title": "Projet-AGL API",
    "description": "The API documentation for Projet-AGL backend",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
...
      "UpdateParcoursDto": {
        "type": "object",
        "properties": {
          "parcoursId": {
            "type": "string",
            "description": "ID du parcours choisi",
            "example": "d290f1ee-6c54-4b01-90e6-d701748f0851"
          }
        },
        "required": [
          "parcoursId"
        ]
      },
      "SimpleChallengeCompleteResponseDto": {
        "type": "object",
        "properties": {
          "feuilles": {
            "type": "number",
            "example": 550
          },
          "lastSimpleChallengeCompletedAt": {
            "type": "string",
            "example": "2026-05-12T10:00:00.000Z"
          }
        },
        "required": [
          "feuilles",
          "lastSimpleChallengeCompletedAt"
        ]
      },
      "ScoreHistoryCategoryDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "description": "Identifiant de la catégorie",
            "example": "transport"
          },
          "name": {
            "type": "string",
            "description": "Libellé de la catégorie",
            "example": "Transport"
          },
          "totalKgCo2ePerYear": {
            "type": "number",
            "description": "Total kgCO2e/an pour cette catégorie",
            "example": 2150.4
          }
        },
        "required": [
          "id",
          "name",
          "totalKgCo2ePerYear"
        ]
      },
      "ScoreHistoryResponseDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "description": "ID de l'entrée d'historique",
            "example": "f839d-21cd-40b9-a212-32a2f81903fa"
          },
          "score": {
            "type": "number",
            "description": "Score carbone obtenu",
            "example": 1450.5
          },
          "created_at": {
            "format": "date-time",
            "type": "string",
            "description": "Date de l'enregistrement",
            "example": "2026-03-16T18:30:00.000Z"
          },
          "categories_scores": {
            "description": "Totaux par catégorie au moment de l’enregistrement (peut être null pour les anciennes entrées).",
            "nullable": true,
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ScoreHistoryCategoryDto"
            }
          }
        },
        "required": [
          "id",
          "score",
          "created_at"
        ]
      },
...
    }
  }
}